refactor: cn함수 추가 외 2건 코드 수정 (#65)
- kakao, naver Button cn함수 추가
- api-request-type 폴더 생성후 auth-request-type 파일 이동
- login / logout 함수명 변경
